Renaming and Reconfiguration
Renamed the class mdview to better convey its functionality. Reconfigured the class to be a non-owning view of a std::vector object, henceforth to be referred to as the "target".  The mdview constructor now takes as parameters a reference to the target as well as the extents of multidimensional view. mdview therefore a structured "manipulator" of the elements in the target. Also, the syntax of the interface was improved via the overloaded function call operator: ().